requirements.txt
django>=3.0.7

djangorestframework>=3.11.0

pytest>=5.4.3
pytest-benchmark>=3.2.3
pytest-django>=3.9.0
pytest-mock>=3.1.1
pytest-runner>=5.2

pylint-django>=2.0.15